+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## [1.1.1] - 2022-24-01
+
+### Changed
+- Changed notification layouts for discordproxy
+
+### Fixed
+- Fixed contracts showing donation icons only if contract had no donations.
+- Fixees #25, fixed issue with notifications when an alt corp was used as manager
+- Fixed rejected notifications not going out to sellers
+
 ## [1.1.0] - 2022-22-01
 
 ### Added